refactor: Move fileSizeFilter to its own file
